+++ src/frontend/ui/parametrage-utilisateurs/EditUser/EditUser.tsx
@@ -14,6 +14,7 @@ import { InstitutionModel } from "../../../../../database/models/InstitutionMode
 import { ProfilModel } from "../../../../../database/models/ProfilModel";
 import { RoleModel } from "../../../../../database/models/RoleModel";
 import { UtilisateurModel } from "../../../../../database/models/UtilisateurModel";
+import { Role, RoleLabel } from "../../../../commons/Role";
 import { formatDateAndHours } from "../../../utils/dateUtils";
 import { useDependencies } from "../../commun/contexts/useDependencies";
 import { useBreadcrumb } from "../../commun/hooks/useBreadcrumb";
@@ -125,15 +126,19 @@ export const EditUser = ({ user, institutions, profiles, roles }: UsersListPageP
       redirectPage("/settings/users?status=edit_successfully");
     });
   }
-  //only "Admin national" can update it self || Admin regional cant update, delete, to (Admin National)
+
   const pageDetails =
-    (data?.user?.idUser === user.code && data?.user?.role !== 1) || ((data?.user?.role as number) > Number.parseInt(user.roleId) && data?.user?.idUser !== user.code);
+    // Seul un Administrateur national peut se modifier lui même
+    (data?.user?.idUser === user.code && data?.user?.role !== Role.ADMIN_NAT) ||
+    // Un adminitrateur regional ne peut pas modifier d’administrateur national ou d’administrateur central
+    ((data?.user?.role as number) === Role.ADMIN_REG && (Number.parseInt(user.roleId) === Role.ADMIN_NAT || Number.parseInt(user.roleId) === Role.ADMIN_CENTR));
 
   let rolesF;
-  if (data?.user?.role === 1) {
+  if (data?.user?.role === Role.ADMIN_NAT) {
     rolesF = roles;
   } else {
-    rolesF = roles.filter((obj) => obj.code !== "ADMIN_NAT");
+    // Seul les admin nationaux peuvent donner des droits d’admin national et central à un utilisateur
+    rolesF = roles.filter((obj) => obj.code !== RoleLabel.ADMIN_NAT && obj.code !== RoleLabel.ADMIN_CENTR);
   }
 
   return (
@@ -266,7 +271,7 @@ export const EditUser = ({ user, institutions, profiles, roles }: UsersListPageP
                             aria-describedby={`checkboxes-${item.code}-messages`}
                             checked={userInfo.profiles && userInfo.profiles.includes(item.code)}
                             className={`${styles["input--checkbox--error"]} `}
-                            disabled={pageDetails || (item.label === "Consultation administrateur national" && data?.user?.role !== 1)}
+                            disabled={pageDetails || (item.label === "Consultation administrateur national" && data?.user?.role !== Role.ADMIN_NAT)}
                             id={`${item.code}`}
                             name="profiles"
                             onChange={handleChange}